Changes to be committed: 	modified:   include/Sensors.h 	modified:   include/debugging.h 	modified:   src/DataBackupSD.cpp 	modified:   src/main.cc

+++ include/Sensors.h
@@ -177,7 +177,7 @@ void Sensors<Adafruit_ADS1115>::get_data(Data& data) {
 template<>
 inline
 void Sensors<AS5048A>::get_data(Data& data) {
-    data.windDirection = (int) s->getRotationInDegrees();
+    data.windDirection = (double) s->getRotationInDegrees();
 }
 
 template<>
